create another js file using modularity
