+++ public/index.html
@@ -16,6 +16,6 @@
       <video src="/rain.mp4" class="video" playsinline></video>
     </div>
 
-    <script src="../src/script.js"></script>
+    <script src="../src/script.js" type="module" ></script>
   </body>
 </html>
